// src/generated/visualThemeBatch2.ts
export const VISUAL_THEME_BATCH_2_THEME_IDS = [
  "blues",
  "pop",
  "ska",
  "rocksteady",
  "soul",
  "punk",
  "1940s",
  "1950s",
  "1960s",
  "1970s",
  "90s-rave",
  "hip-hop",
  "greek",
  "french",
  "italian"
] as const